backend/requirements.txt
torch
torchvision
git+https://github.com/openai/CLIP.git
faiss-cpu
streamlit
pandas
pillow
scikit-learn
numpy
transformers
opencv-python
sentencepiece
tiktoken
protobuf
rank_bm25
openpyxl
fastapi
uvicorn
python-multipart
openai
python-dotenv


